+++ auth/metamask/src/metamask.module.ts
@@ -1,12 +1,9 @@
 import { Module } from "@nestjs/common";
 
-import { LicenseModule, licenseProvider } from "@ethberry/nest-js-module-license";
-
 import { MetamaskService } from "./metamask.service";
 
 @Module({
-  imports: [LicenseModule.deferred()],
-  providers: [licenseProvider, MetamaskService],
+  providers: [MetamaskService],
   exports: [MetamaskService],
 })
 export class MetamaskModule {}