+++ auth/particle/src/particle.module.ts
@@ -1,12 +1,9 @@
 import { Module } from "@nestjs/common";
 
-import { LicenseModule, licenseProvider } from "@ethberry/nest-js-module-license";
-
 import { ParticleService } from "./particle.service";
 
 @Module({
-  imports: [LicenseModule.deferred()],
-  providers: [licenseProvider, ParticleService],
+  providers: [ParticleService],
   exports: [ParticleService],
 })
 export class ParticleModule {}