+++ mail/mailjet/src/mailjet.module.ts
@@ -1,15 +1,12 @@
 import { DynamicModule, Logger, Module } from "@nestjs/common";
 import { createConfigurableDynamicRootModule } from "@golevelup/nestjs-modules";
 
-import { LicenseModule, licenseProvider } from "@ethberry/nest-js-module-license";
-
 import { MAILJET_OPTIONS_PROVIDER } from "./mailjet.constants";
 import { MailjetService } from "./mailjet.service";
 import { IMailjetOptions } from "./interfaces";
 
 @Module({
-  imports: [LicenseModule.deferred()],
-  providers: [licenseProvider, Logger, MailjetService],
+  providers: [Logger, MailjetService],
   exports: [MailjetService],
 })
 export class MailjetModule extends createConfigurableDynamicRootModule<MailjetModule, IMailjetOptions>(