+++ modules/gee-test/src/geetest.module.ts
@@ -2,16 +2,14 @@ import { Module, Logger } from "@nestjs/common";
 import { HttpModule } from "@nestjs/axios";
 import { ConfigModule } from "@nestjs/config";
 
-import { LicenseModule, licenseProvider } from "@ethberry/nest-js-module-license";
-
 import { GeeTestController } from "./geetest.controller";
 import { GeeTestService } from "./geetest.service";
 import { ValidateGeeTest } from "./geetest.validator";
 
 @Module({
-  imports: [LicenseModule.deferred(), ConfigModule, HttpModule],
+  imports: [ConfigModule, HttpModule],
   controllers: [GeeTestController],
-  providers: [licenseProvider, Logger, GeeTestService, ValidateGeeTest],
+  providers: [Logger, GeeTestService, ValidateGeeTest],
   exports: [GeeTestService, ValidateGeeTest],
 })
 export class GeeTestModule {}