+++ modules/s3/src/s3.module.ts
@@ -1,15 +1,12 @@
 import { DynamicModule, Logger, Module } from "@nestjs/common";
 import { createConfigurableDynamicRootModule } from "@golevelup/nestjs-modules";
 
-import { LicenseModule, licenseProvider } from "@ethberry/nest-js-module-license";
-
 import { IS3Options } from "./interfaces";
 import { S3Service } from "./s3.service";
 import { S3_OPTIONS_PROVIDER } from "./s3.constants";
 
 @Module({
-  imports: [LicenseModule.deferred()],
-  providers: [Logger, licenseProvider, S3Service],
+  providers: [Logger, S3Service],
   exports: [S3Service],
 })
 export class S3Module extends createConfigurableDynamicRootModule<S3Module, IS3Options>(S3_OPTIONS_PROVIDER) {