+++ modules/throttler/src/throttler.module.ts
@@ -5,13 +5,10 @@ import type { ThrottlerModuleOptions } from "@nestjs/throttler";
 import { RedisManager } from "@liaoliaots/nestjs-redis";
 import { ThrottlerStorageRedisService } from "@nest-lab/throttler-storage-redis";
 
-import { LicenseModule, licenseProvider } from "@ethberry/nest-js-module-license";
-
 import { THROTTLE_STORE } from "./throttler.constants";
 
 @Module({
   imports: [
-    LicenseModule.deferred(),
     ThrottlerModule.forRootAsync({
       imports: [ConfigModule],
       inject: [ConfigService, RedisManager],
@@ -26,6 +23,5 @@ import { THROTTLE_STORE } from "./throttler.constants";
       }),
     }),
   ],
-  providers: [licenseProvider],
 })
 export class EthBerryThrottlerModule {}